Create README.md
Used MNIST dataset for handwritten digits and starter code for a single hidden layer, 10 node network. First time working with a neural network so fun to mess around with the code from scratch, adding more layers and nodes and artificially expanding the dataset and other things to experiment with how it worked. Definitely could benefit from using threading and smaller branches in the future.